cli/tests/requirements.txt
pytest==8.3.3
pytest-cov==5.0.0
responses==0.25.3
pyyaml==6.0.2
jsonschema==4.23.0
